refactor(layout_task.rs): move some functions to query.rs
